Move CatalogOperationsTest to openhouse-spark-itest module
