+++ integrations/spark/openhouse-spark-itest/build.gradle
@@ -13,7 +13,7 @@ dependencies {
   }
 
   testImplementation(project(':tables-test-fixtures_2.12'))
-
+  testImplementation 'org.junit.platform:junit-platform-runner:1.8.2'
   // Required to test /tables mockserver
   testImplementation "com.squareup.okhttp3:okhttp:4.0.1"
   testImplementation "com.squareup.okhttp3:mockwebserver:4.0.1"
@@ -34,19 +34,3 @@ sourceSets {
     }
   }
 }
-
-// Specify order of tests to avoid spark session conflicts
-// Non-openhouse catalog tests should run before e2e tests start
-task statementTest(type: Test) {
-  filter {
-    includeTestsMatching 'com.linkedin.openhouse.spark.statementtest.*'
-  }
-}
-
-test {
-  filter {
-    excludeTestsMatching 'com.linkedin.openhouse.spark.statementtest.*'
-  }
-}
-
-test.dependsOn statementTest
