+++ integrations/spark/openhouse-spark-itest/src/test/java/com/linkedin/openhouse/spark/catalogtest/CatalogOperationTest.java
@@ -1,4 +1,4 @@
-package com.linkedin.openhouse.spark;
+package com.linkedin.openhouse.spark.catalogtest;
 
 import com.linkedin.openhouse.tablestest.OpenHouseSparkITest;
 import java.util.HashMap;
